+++ shopapp/src/main/java/com/project/shopapp/responses/ProductResponse.java
@@ -20,7 +20,7 @@ import java.util.List;
 public class ProductResponse extends BaseResponse{
     private Long id;
     private String name;
-    private Float price;
+    private Long price;
     private String thumbnail;
     private String description;
     @JsonProperty("product_images")
